% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy}
\alias{dummy}
\title{Dummy data set documentation}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{carrier}{... add description ...}
\item{name}{... add description ...}
...
}
}
\source{
See data-raw/dummy.R
}
\usage{
dummy
}
\description{
A dataset containing ...
}
\keyword{datasets}
