% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy2}
\alias{dummy2}
\title{More dummy data documentation}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
dummy2
}
\description{
\itemize{
\item \emph{italics}
\item \strong{bold}
\item \code{r_function_call(with = "arguments")}
\item \code{\link{function}}: function in this package.
\item \code{\link[MASS]{abbey}}: function in another package.
\item \link[=dest]{name}: link to dest, but show name.
\item \code{\link[MASS:abbey]{name}}: link to function in another package, but show name.
\item \linkS4class{abc}: link to an S4 class.
\item \url{https://fzstat-rnaseq-marray.netlify.app/}: a url.
\item \href{https://keepyourheadup.netlify.app/}{Keep Your Head Up!}: a url with custom link text.
\item \email{francois.collin@gmail.com} (note the doubled @): an email address.
}
}
\details{
\preformatted{The following function turns an R data frame into the correct format. 
It ignores column and row names, but should get you started.

tabular <- function(df, ...) {
stopifnot(is.data.frame(df))

align <- function(x) if (is.numeric(x)) "r" else "l"
col_align <- vapply(df, align, character(1))

cols <- lapply(df, format, ...)
contents <- do.call("paste",
 c(cols, list(sep = " \\tab ", collapse = "\\cr\n  ")))

paste("\\tabular{", paste(col_align, collapse = ""), "}{\n  ",
 contents, "\n}\n", sep = "")
}

cat(tabular(mtcars[1:5, 1:5]))
}
\tabular{rrrrr}{
21.0 \tab 6 \tab 160 \tab 110 \tab 3.90\cr
21.0 \tab 6 \tab 160 \tab 110 \tab 3.90\cr
22.8 \tab 4 \tab 108 \tab  93 \tab 3.85\cr
21.4 \tab 6 \tab 258 \tab 110 \tab 3.08\cr
18.7 \tab 8 \tab 360 \tab 175 \tab 3.15
}
}
\keyword{datasets}
