% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFns.R
\name{file_management}
\alias{file_management}
\alias{obj_save}
\alias{obj_load}
\title{Saving to and loading from the file system}
\usage{
obj_save(file_name = "", obj_name = "", data_dir = file.path("data"))

obj_load(file_name = "", obj_name = "", data_dir = file.path("data"))
}
\arguments{
\item{file_name}{A character string file name.}

\item{obj_name}{A character string referring to an object.}

\item{data_dir}{A path to a system folder.}
}
\description{
Saving to and loading from the file system
}
\examples{
\dontrun{
# Write object to file system
samples_frm <- data.table::fread("path_to_raw_data_file")
obj_save('thisProject_samples_frm', 'samples_frm', data_dir='data')

# load object from file system
obj_load('thisProject_samples_frm', 'samples_frm', data_dir='data')
head(samples_frm)
...
}
}
