% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicScaleR.R
\docType{package}
\name{omicScaleR}
\alias{omicScaleR}
\title{omicScaleR: A package for Omic scale data analysis}
\description{
The foo package provides three categories of functions:
quality assessment, model building, model dignostic and evaluation
The omicScaleR functions ...
}
